backend/requirements.txt
langchain
langchain-openai
langchain-community
langgraph
pydantic
python-dotenv
ddgs